backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
scikit-learn==1.3.1
numpy==1.24.3
openai==1.3.0
huggingface-hub==0.19.0
gunicorn==21.2.0
